+++ packages/zp-data/src/mongodb/Service.ts
@@ -5,7 +5,7 @@ import { Id, NullableId, Paginated, Params, Service } from 'zapnode'
 import { Collection } from 'mongodb'
 import MongoDBRepository from './Repository'
 
-class MongoDBService<Result = any, Data = Partial<Result>, ServiceParams extends Params<any> = any> implements Service<Result, Data, ServiceParams> {
+class MongoDBService<Result = any, Data = any, ServiceParams extends Params = any> implements Service<Result, Data, ServiceParams> {
   protected repository: MongoDBRepository<Result, Data, ServiceParams>
 
   constructor (collection: Collection) {
@@ -16,13 +16,11 @@ class MongoDBService<Result = any, Data = Partial<Result>, ServiceParams extends
     this.repository = new MongoDBRepository<Result, Data, ServiceParams>(collection)
   }
 
-  async find (params?: ServiceParams): Promise<Result[]>
-  async find (params?: ServiceParams & { query: { $paginate: true } }): Promise<Paginated<Result>>
   async find (params?: ServiceParams): Promise<Paginated<Result> | Result[]> {
     return this.repository.find(params)
   }
 
-  async get (id: Id, params: ServiceParams = {} as ServiceParams): Promise<Result> {
+  async get (id: Id, params?: ServiceParams): Promise<Result> {
     return this.repository.get(id, params)
   }
 
@@ -30,20 +28,14 @@ class MongoDBService<Result = any, Data = Partial<Result>, ServiceParams extends
     return this.repository.update(id, data, params)
   }
 
-  async create (data: Data, params?: ServiceParams): Promise<Result>
-  async create (data: Data[], params?: ServiceParams): Promise<Result[]>
-  async create (data: Data | Data[], params: ServiceParams = {} as ServiceParams): Promise<Result | Result[]> {
+  async create (data: Data | Data[], params?: ServiceParams): Promise<Result | Result[]> {
     return this.repository.create(data, params)
   }
 
-  async patch (id: null, data: Data, params?: ServiceParams): Promise<Result[]>
-  async patch (id: Id, data: Data, params?: ServiceParams): Promise<Result>
   async patch (id: NullableId, data: Data, params?: ServiceParams): Promise<Result | Result[]> {
     return this.repository.patch(id, data, params)
   }
 
-  async remove (id: null, params?: ServiceParams): Promise<Result[]>
-  async remove (id: Id, params?: ServiceParams): Promise<Result>
   async remove (id: NullableId, params?: ServiceParams): Promise<Result | Result[]> {
     return this.repository.remove(id, params)
   }