+++ packages/zp-auth/CHANGELOG.md
@@ -1,5 +1,14 @@
 # zapnode-auth
 
+## 1.0.8
+
+### Patch Changes
+
+- Update MongoDB service and Repository
+- Updated dependencies
+  - zapnode@1.0.8
+  - zapnode-plugins@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes