+++ packages/zp-plugins/CHANGELOG.md
@@ -1,5 +1,13 @@
 # zapnode-plugins
 
+## 1.0.8
+
+### Patch Changes
+
+- Update MongoDB service and Repository
+- Updated dependencies
+  - zapnode@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes